--- project/src/main/java/ict/project/resume/controller/UserController.java
@@ -1,22 +1,0 @@
-package ict.project.resume.controller;
-
-import ict.project.resume.entity.UserEntity;
-import ict.project.resume.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/users")
-public class UserController {
-
-    private final UserRepository userRepository;
-
-    @GetMapping("/{id}")
-    public ResponseEntity<?> get(@PathVariable Integer id) {
-        return userRepository.findById(id)
-                .<ResponseEntity<?>>map(ResponseEntity::ok)
-                .orElseGet(() -> ResponseEntity.notFound().build());
-    }
-}
